move create_game/1 to game_server
